+++ packages/sdks/snippets/sveltekit/src/routes/announcements/[...catchall]/+page.svelte
@@ -3,11 +3,14 @@
 <!-- src/routes/announcements/[...catchall]/+page.svelte -->
 
 <script>
-import { isPreviewing, Content } from "@builder.io/sdk-svelte";
+    import { isPreviewing, Content } from "@builder.io/sdk-svelte";
 
-export let data;
+    const apiKey = "ee9f13b4981e489a9a1209887695ef2b"
+    const model = "announcement-bar"
 
-const canShowContent = data.content || isPreviewing();
+    export let data;
+
+    const canShowContent = data.content || isPreviewing();
 </script>
 
 <main>
@@ -15,7 +18,7 @@ const canShowContent = data.content || isPreviewing();
     <h2>Below is your Builder Content:</h2>
     {#if canShowContent}
         <div>page Title: {data.content?.data?.title || "Unpublished"}</div>
-        <Content model="announcement-bar" content={data.content} apiKey={'ee9f13b4981e489a9a1209887695ef2b'} />
+        <Content model={model} content={data.content} apiKey={apiKey} />
     {:else}
         <div>Announcement Bar not Found</div>
     {/if}
